refactor: split app by layout & refactor chakra theme provider
