+++ src/app/(legacy-layout)/about/page.tsx
@@ -1,0 +1,22 @@
+import { Metadata } from 'next'
+
+import * as components from '@/components'
+
+export const metadata: Metadata = {
+    title: 'Peanut Protocol',
+    description: 'Text Tokens',
+    metadataBase: new URL('https://peanut.to'),
+    icons: {
+        icon: '/logo-favicon.png',
+    },
+    openGraph: {
+        images: [
+            {
+                url: '/metadata-img.png',
+            },
+        ],
+    },
+}
+export default function AboutPage() {
+    return <components.About />
+}
