+++ src/app/(legacy-layout)/blog/page.tsx
@@ -1,0 +1,24 @@
+import { Metadata } from 'next'
+
+import * as components from '@/components'
+
+export const metadata: Metadata = {
+    title: 'Peanut Protocol',
+    description: 'Text Tokens',
+    metadataBase: new URL('https://peanut.to'),
+
+    icons: {
+        icon: '/logo-favicon.png',
+    },
+    openGraph: {
+        images: [
+            {
+                url: '/metadata-img.png',
+            },
+        ],
+    },
+}
+
+export default function BlogPage() {
+    return <components.Blog />
+}
