+++ src/app/(legacy-layout)/cashout/status/page.tsx
@@ -1,0 +1,9 @@
+import * as components from '@/components'
+
+export const dynamic = 'force-dynamic'
+
+// TODO: add metadata
+
+export default function CashoutStatusPage() {
+    return <components.CashoutStatus />
+}
