+++ src/app/(legacy-layout)/kyc/page.tsx
@@ -1,0 +1,24 @@
+import * as components from '@/components'
+
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+    title: 'Peanut Protocol',
+    description: 'Send crypto via link',
+    metadataBase: new URL('https://peanut.to'),
+
+    icons: {
+        icon: '/logo-favicon.png',
+    },
+    openGraph: {
+        images: [
+            {
+                url: '/metadata-img.png',
+            },
+        ],
+    },
+}
+
+export default function KycPage() {
+    return <components.KYCComponent />
+}
