+++ src/app/(legacy-layout)/profile/page.tsx
@@ -1,0 +1,24 @@
+import * as components from '@/components'
+
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+    title: 'Peanut Protocol',
+    description: 'Text Tokens',
+    metadataBase: new URL('https://peanut.to'),
+
+    icons: {
+        icon: '/logo-favicon.png',
+    },
+    openGraph: {
+        images: [
+            {
+                url: '/metadata-img.png',
+            },
+        ],
+    },
+}
+
+export default function ProfilePage() {
+    return <components.Profile />
+}
