--- src/app/cashout/page.tsx
@@ -1,14 +1,0 @@
-import * as components from '@/components'
-import Layout from '@/components/Global/Layout'
-
-export const dynamic = 'force-dynamic'
-
-// TODO: add metadata
-
-export default function CashoutPage() {
-    return (
-        <Layout>
-            <components.Cashout />
-        </Layout>
-    )
-}
