+++ src/config/index.ts
@@ -1,2 +1,3 @@
 export * from './wagmi.config'
 export * from './peanut.config'
+export * from './theme.config'
