+++ src/config/peanut.config.tsx
@@ -11,9 +11,6 @@ import '../../sentry.client.config'
 import '../../sentry.server.config'
 import '../../sentry.edge.config'
 import 'react-tooltip/dist/react-tooltip.css'
-// import LogRocket from 'logrocket'
-import { ChakraProvider } from '@chakra-ui/react'
-import * as styles from '@/styles/theme'
 
 export function PeanutProvider({ children }: { children: React.ReactNode }) {
     useEffect(() => {
@@ -27,7 +24,7 @@ export function PeanutProvider({ children }: { children: React.ReactNode }) {
 
     return (
         <config.ContextProvider>
-            <ChakraProvider theme={styles.theme}> {children}</ChakraProvider>
+            {children}
             <Analytics />
         </config.ContextProvider>
     )
